# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Synchrome API",
        "version": "0.1"
    },
    "paths": {
...
        "/calendars": {
            "get": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Fetch list of calendars.",
                "operationId": "calendarsIndex",
                "responses": {
                    "200": {
                        "description": "List of calendars."
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Create a new calendar.",
                "operationId": "calendarsStore",
                "parameters": [
                    {
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateCalendarRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/calendars/{id}": {
            "get": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Fetch list of calendars.",
                "operationId": "calendarsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A calendar."
                    },
                    "404": {
                        "description": "Not found."
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Delete a calendar.",
                "operationId": "calendarsDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted."
                    },
                    "404": {
                        "description": "Not found."
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Update a calendar.",
                "operationId": "calendarsUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateCalendarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/calendars/{id}/events": {
            "get": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Fetch list of events of a calendar.",
                "operationId": "calendarsGetEvents",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Events list."
                    },
                    "404": {
                        "description": "Not found."
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Create a new event of a calendar.",
                "operationId": "calendarsEventStore",
                "parameters": [
                    {
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateCalendarEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/calendars/{calendar_id}/events/{event_id}": {
            "delete": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Delete a calendar.",
                "operationId": "calendarsEventDestroy",
                "parameters": [
                    {
                        "name": "calendar_id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "event_id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted."
                    },
                    "404": {
                        "description": "Not found."
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Calendars"
                ],
                "summary": "Update an event of a calendar.",
                "operationId": "calendarsEventUpdate",
                "parameters": [
                    {
                        "name": "calendar_id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "event_id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateCalendarEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "CreateCalendarRequest": {
            "properties": {
                "name": {
                    "description": "Class Calendar",
                    "type": "string"
                },
                "start": {
                    "description": "Class Calendar",
                    "type": "string"
                },
                "end": {
                    "description": "Class Calendar",
                    "type": "string"
                },
                "events": {
                    "description": "Class Calendar",
                    "type": "array",
                    "items": {
                        "properties": {
                            "name": {
                                "description": "Class Calendar",
                                "type": "string"
                            },
                            "start": {
                                "description": "Class Calendar",
                                "type": "string"
                            },
                            "end": {
                                "description": "Class Calendar",
                                "type": "string"
                            },
                            "attendanceTypeId": {
                                "description": "Class Calendar",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                }
            }
        },
        "UpdateCalendarRequest": {
            "properties": {
                "name": {
                    "description": "Class Calendar",
                    "type": "string"
                },
                "start": {
                    "description": "Class Calendar",
                    "type": "string"
                },
                "end": {
                    "description": "Class Calendar",
                    "type": "string"
                }
            }
        },
        "CreateCalendarEventRequest": {
            "properties": {
                "name": {
                    "description": "Class Calendar",
                    "type": "string"
                },
                "start": {
                    "description": "Class Calendar",
                    "type": "string"
                },
                "end": {
                    "description": "Class Calendar",
                    "type": "string"
                },
                "attendanceTypeId": {
                    "description": "Class Calendar",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "basicAuth": {
            "type": "basic",
            "description": "Authenticates via HTTP Basic",
            "name": "authorization"
        }
    }
}